rework sentry middleware
