+++ packages/store-indexer/src/koa-middleware/compress.ts
