+++ packages/store-indexer/src/koa-middleware/healthcheck.ts
